requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
google-generativeai==0.3.2
qrcode==7.4.2
Pillow==10.2.0
python-dateutil==2.8.2